Updated to pass demographics test.

+++ ehr_functions/utils/numerics.py
@@ -1,3 +1,6 @@
+import numpy as np
+
+
 def make_numeric(df, cols=None, ignore_cols=None, return_new_cols=False):
     """
     :param df:
@@ -21,13 +24,13 @@ def make_numeric(df, cols=None, ignore_cols=None, return_new_cols=False):
 
             # If less than two then just convert to be boolean
             if len(unique_values) <= 2:
-                df[col] = df[col].apply(lambda x: True if x == unique_values[0] else False)
+                df[col] = df[col].apply(lambda x: True if x == unique_values[0] else False).astype(np.int64)
             else:
                 # Otherwise make separate columns for each unique value
                 for value in unique_values:
                     if value != '':
                         new_col_name = col + '_' + value
-                        df[new_col_name] = df[col].apply(lambda x: True if x == value else False)
+                        df[new_col_name] = df[col].apply(lambda x: True if x == value else False).astype(np.int64)
                         new_cols.append(new_col_name)
 
                 df = df.drop(columns=[col])
